format all code

+++ adaptations/train_mpa.py
@@ -1,6 +1,8 @@
 import glob
+import os
 import shutil
-import os, sys
+import sys
+
 os.chdir(sys.path[0])
 import torch
 import torch.nn as nn
@@ -12,7 +14,6 @@ import torch.optim as optim
 import torch.backends.cudnn as cudnn
 import torch.nn.functional as F
 import logging, sys
-import time
 from tensorboardX import SummaryWriter
 from model.trans4pass import Trans4PASS_v1, Trans4PASS_v2
 from model.discriminator import FCDiscriminator
@@ -47,7 +48,7 @@ LEARNING_RATE = 2.5e-6
 MOMENTUM = 0.9
 NUM_CLASSES = 19
 NUM_STEPS = 100000
-NUM_STEPS_STOP = 80000 # early stopping
+NUM_STEPS_STOP = 80000  # early stopping
 NUM_PROTOTYPE = 50
 POWER = 0.9
 RANDOM_SEED = 1234
@@ -94,6 +95,7 @@ NAME_CLASSES = [
     "train",
     "motocycle",
     "bicycle"]
+
 
 def get_arguments():
     """Parse all the arguments provided from the CLI.
@@ -180,7 +182,9 @@ def get_arguments():
                         help="Path to save result.")
     return parser.parse_args()
 
+
 args = get_arguments()
+
 
 def setup_logger(name, save_dir, filename="log.txt", mode='w'):
     logging.root.name = name
@@ -200,7 +204,9 @@ def setup_logger(name, save_dir, filename="log.txt", mode='w'):
     ch.setFormatter(formatter)
     logging.root.addHandler(ch)
 
+
 setup_logger('Trans4PASS', SNAPSHOT_DIR)
+
 
 def lr_poly(base_lr, iter, max_iter, power):
     return base_lr * ((1 - float(iter) / max_iter) ** (power))
@@ -236,6 +242,7 @@ def load_my_state_dict(model, state_dict):  # custom function to load model when
         else:
             own_state[name].copy_(param)
     return model
+
 
 def main():
     """Create the model and start the training."""
@@ -283,12 +290,14 @@ def main():
 
     init_mem_j_path = 'init_memory_joint_ms.npy'
     if not os.path.exists(init_mem_j_path):
-        trainset_temp = densepassDataSet(args.data_dir_target, args.data_list_target, crop_size=input_size_target, set='train',
+        trainset_temp = densepassDataSet(args.data_dir_target, args.data_list_target, crop_size=input_size_target,
+                                         set='train',
                                          ssl_dir=SSL_DIR)
         trainloader_temp = data.DataLoader(trainset_temp, batch_size=1, shuffle=False)
         testset_temp = CSSrcDataSet(args.data_dir, args.data_list, crop_size=input_size, set='train')
         testloader_temp = data.DataLoader(testset_temp, batch_size=1, shuffle=False)
-        init_mem = init_memory(trainloader_temp, testloader_temp, model, num_classes=args.num_classes, save_path=init_mem_j_path)
+        init_mem = init_memory(trainloader_temp, testloader_temp, model, num_classes=args.num_classes,
+                               save_path=init_mem_j_path)
         del trainloader_temp, trainset_temp, testset_temp, testloader_temp
     else:
         init_mem = np.load(init_mem_j_path)
@@ -309,13 +318,16 @@ def main():
 
     # init data loader
     trainset = CSSrcDataSet(args.data_dir, args.data_list, max_iters=args.num_steps * args.iter_size * args.batch_size,
-                            crop_size=input_size, scale=args.random_scale, mirror=args.random_mirror, mean=IMG_MEAN, set=args.set)
+                            crop_size=input_size, scale=args.random_scale, mirror=args.random_mirror, mean=IMG_MEAN,
+                            set=args.set)
     trainloader = data.DataLoader(trainset, batch_size=args.batch_size, shuffle=True, num_workers=args.num_workers,
                                   pin_memory=True)
     trainloader_iter = enumerate(trainloader)
     # --- SSL_DIR
-    targetset = densepassDataSet(args.data_dir_target, args.data_list_target, max_iters=args.num_steps * args.iter_size * args.batch_size,
-                                 crop_size=input_size_target, scale=False, mirror=args.random_mirror, mean=IMG_MEAN, set=args.set,
+    targetset = densepassDataSet(args.data_dir_target, args.data_list_target,
+                                 max_iters=args.num_steps * args.iter_size * args.batch_size,
+                                 crop_size=input_size_target, scale=False, mirror=args.random_mirror, mean=IMG_MEAN,
+                                 set=args.set,
                                  ssl_dir=SSL_DIR, trans=TARGET_TRANSFORM)
     targetloader = data.DataLoader(targetset, batch_size=args.batch_size, shuffle=True, num_workers=args.num_workers,
                                    pin_memory=True)
@@ -332,8 +344,8 @@ def main():
 
     model.train()
     # init optimizer
-    optimizer = optim.SGD(model.optim_parameters(args), 
-                        lr=args.learning_rate, momentum=args.momentum, weight_decay=args.weight_decay)
+    optimizer = optim.SGD(model.optim_parameters(args),
+                          lr=args.learning_rate, momentum=args.momentum, weight_decay=args.weight_decay)
     optimizer.zero_grad()
 
     optimizer_D = optim.Adam(model_D.parameters(), lr=args.learning_rate_D, betas=(0.9, 0.99))
@@ -376,7 +388,6 @@ def main():
         if not os.path.exists(args.log_dir):
             os.makedirs(args.log_dir)
         writer = SummaryWriter(args.log_dir)
-
 
     # start training
     for i_iter in range(Iter, args.num_steps):
@@ -471,9 +482,10 @@ def main():
                 for clsid in range(NUM_CLASSES):
                     if len(init_batch_mem[clsid]) > 1:
                         batch_center = np.mean(init_batch_mem[clsid])[None, ...]
-                        cluster_batch_feats_mem_s[clsid] = torch.from_numpy(batch_center).to(cluster_batch_feats_mem_s.dtype)
+                        cluster_batch_feats_mem_s[clsid] = torch.from_numpy(batch_center).to(
+                            cluster_batch_feats_mem_s.dtype)
                 # ema
-                init_mem = init_mem * MOMENTUM_MEM + cluster_batch_feats_mem_s * (1-MOMENTUM_MEM)
+                init_mem = init_mem * MOMENTUM_MEM + cluster_batch_feats_mem_s * (1 - MOMENTUM_MEM)
 
         optimizer.step()
         optimizer_D.step()
@@ -492,8 +504,10 @@ def main():
                 for key, val in scalar_info.items():
                     writer.add_scalar(key, val, i_iter)
         if i_iter % 10 == 0:
-            logging.info('iter={0:8d}/{1:8d}, l_seg={2:.3f}, l_seg_t={7:.3f}, l_adv={3:.3f} l_D={4:.3f}, l_kl_s={5:.3f}, l_kl_t={6:.3f}'.format(
-                  i_iter, args.num_steps, loss_seg_value, loss_adv_target_value, loss_D_value, loss_kl_s_value, loss_kl_t_value, loss_seg_value_t))
+            logging.info(
+                'iter={0:8d}/{1:8d}, l_seg={2:.3f}, l_seg_t={7:.3f}, l_adv={3:.3f} l_D={4:.3f}, l_kl_s={5:.3f}, l_kl_t={6:.3f}'.format(
+                    i_iter, args.num_steps, loss_seg_value, loss_adv_target_value, loss_D_value, loss_kl_s_value,
+                    loss_kl_t_value, loss_seg_value_t))
 
         if i_iter >= args.num_steps_stop - 1:
             logging.info('save model ...')
@@ -512,7 +526,7 @@ def main():
                 with torch.no_grad():
                     output1, output2 = model(Variable(image).to(device))
                 output = output2.cpu().data[0].numpy()
-                output = output.transpose(1,2,0)
+                output = output.transpose(1, 2, 0)
                 output = np.asarray(np.argmax(output, axis=2), dtype=np.uint8)
                 label = label.cpu().data[0].numpy()
                 hist += fast_hist(label.flatten(), output.flatten(), args.num_classes)
